add AwsClientProperties class for aws creds provider
